configure freertos to use newlib allocator
The intent of this change is so that freertos doesn't allocate
17k of ram for its own heap that isn't shared with userland
allocations.

There is an air of Chesterton's Fence about this change, in that I
don't know what the downsides are. I speculate that freertos
maintainers (reasonably) want not to add dependencies on a C library
that they don't control, and that "conventional" embedded applications
don't make as much use of malloc as a Lua interpreter does.

We can't do it the other way around and use the FreeRTOS allocator for
everything, as it doesn't support realloc()

With this change, sbrk(0)=0x20009040 at (down from 2000BAB0)
after all the lua bytecode is loaded